% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn.R
\name{layer}
\alias{layer}
\title{Linear layer}
\usage{
layer(nin, nout, ...)
}
\arguments{
\item{nin}{Number of input features}

\item{nout}{Number of output features}

\item{...}{Passed on to \code{neuron} (\code{act} is the only parameter)}
}
\value{
A list with two elements: \code{params} and \code{forward}.
\itemize{
\item \code{params} is a list with the initial parameters of the model using a default
initialization strategy.
\item \code{forward} is a function that takes a list of \code{nin} \code{Value} and the \code{params} objects and
returns a list of \code{nout} \code{Value} objects. Essentially doing what a \code{torch::nn_linear} layer does.
}
}
\description{
Linear layer
}
