% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn.R
\name{mlp}
\alias{mlp}
\title{Creates a Multi Layer Perceptron model}
\usage{
mlp(nin, nouts)
}
\arguments{
\item{nin}{Number of input features}

\item{nouts}{An integer vector with the number of neurons in each layer}
}
\value{
A list with two elements: \code{params} and \code{forward}.
\itemize{
\item \code{params} is a list with the initial parameters of the model using a default
initialization strategy.
\item \code{forward} is a function that takes a list of \code{nin} \code{Value} and the \code{params} objects and
returns a list of (\code{nouts[length(nouts)]}) \code{Value} objects.
}
}
\description{
Creates a Multi Layer Perceptron model
}
