% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moons.R
\name{moon}
\alias{moon}
\title{Creates a small dataset for experimentation}
\usage{
moon(n, noise = 0.1)
}
\arguments{
\item{n}{Number of samples}

\item{noise}{Standard deviation of the noise}
}
\description{
Creates a small dataset for experimentation
}
