% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{Value}
\alias{Value}
\title{A class that represents a value and how to compute it's gradients}
\usage{
Value(data = NA_real_, grads = NULL, parents = NULL, id = uuid::UUIDgenerate())
}
\arguments{
\item{data}{The value of the object}

\item{grads}{The gradients of the object w.r.t. it's parents}

\item{parents}{The parents of the object}

\item{id}{A unique identifier for the object}
}
\description{
A class that represents a value and how to compute it's gradients
}
