% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{value_and_grad}
\alias{value_and_grad}
\title{Creates a function that can compute the value and gradients of an input function}
\usage{
value_and_grad(f)
}
\arguments{
\item{f}{A function that takes \code{Value} objects as inputs and returns a single \code{Value}
object.}
}
\value{
A function with the same arguments as \code{f} that returns a list with two
elements: \code{value} and \code{grads}. \code{value} is the result of \code{f} and \code{grads} is a named
list with the gradients of the inputs w.r.t. the output of \code{f}.
}
\description{
Creates a function that can compute the value and gradients of an input function
}
